added a function that clears the error message in some cases
when shown to the user the error message about typing just one number and clicking on the equal_button, and when the same user tried typing one new number, the error message was not deleted. So, I implemented one function that solve the problem, immediately deleting the message before the new number is entered.